+++ ployst/core/repos/views.py
@@ -1,6 +1,6 @@
 from rest_framework.viewsets import ModelViewSet
 
-from ployst.apibase.mixins import PermissionsViewSetMixin
+from ployst.core.accounts.mixins import PermissionsViewSetMixin
 
 from .models import Repository, Branch
 from .serializers import RepositorySerializer
